requirements.txt
# 动作识别项目依赖

# 核心依赖
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
paho-mqtt>=1.6.0

# 可选依赖
# pillow>=10.0.0  # 图像处理
# scipy>=1.11.0   # 科学计算
